EXT-1664 male avatars have female bodies
Previous patch improved the situation but reversed it - it was still possible to have a mismatch between an avatar's head and body for their gender.
This patch makes the copying of the gender from your shape wearable to your avatar more explicit, which appears to fix the issue for now.
This should get the behavior correct for testing purposes, and we will look into fixing more underlying issues for the next cycle.

Code reviewed by Seraph