EXT-1784 Server getting spammed whenever you put on a wearable
Made an adjustment to ignore remove requests on items that don't exist, to avoid the extra messages to the server.
Also made a couple of adjustments to ensure removal of alpha masks and tattoos along with other wearable types.
Added these code points to the wiki page detailing what modifications need to be made when adding a new wearable type:
https://wiki.lindenlab.com/wiki/User:Vir#Adding_new_wearables

Code reviewed by Seraph.