DEV-39999: Removed all legacy help buttons and notifications from the code. All help should now go through the new help server, via the Help Browser floater. There should no longer be any custom "?" buttons in the UI - all <button> and <notification> XUI references relating to custom help buttons have been expunged.
(I previously extracted all of the help strings from notifications.xml
and passed these to the doc team to include in the help server pages.)

I have checked that none of the XUI elements that are removed here are
still referenced by any C++ code. I've also confirmed that there are
no new notifications removed that I have not already passed on to the
doc team. And I've also done a pass through the UI to make sure that
all of the floaters still look as they should. Finally, I've confirmed
that the new help system for floaters and side panels still works.

2,276 lines deleted - 0 lines added.