Fix unposted bug: potential incorrect contract in LLMediaEntry::asLLSD(LLSD& sd) Review #26
Callum found this potential bug in LLMediaEntry::asLLSD(LLSD&). It
does not erase the incoming LLSD's WHITELIST_KEY. Every other key is
overwritten in this manner, but for the WHITELIST_KEY it is just
appended. Note that this is in code common to the server...but I do
not believe at this point that there is a bug exhibited by this.

Added unit test to test for this case.