DEV-41317 DEV-42311: Added ad-hoc voice call support.
You can now multiple select users in the People panel, hit Call, and
start an ad-hoc voice conference call with those users.

The most difficult part here was automatically starting the call once
the conference chat panel popped up. We have to wait for the panel to
initialize before we can start a call, so I added another callback to
LLIMModel to enable us to get notified when the panel has initialized.
This is all wrapped up behind a new LLAvatarActions::startAdhocCall()
API.