EXT-5105 Incorrect keyboard shortcut for Inventory in main menu EXT-5334 CTRL+SHIFT+I now brings up InventorySP instead of InventoryFloater
Fixed a regression due to EXT-5105 where CTRL+SHIFT+I was bringing up the InventorySP instead of InventoryFloater.  Note that this reverts some code checked in for EXT-5105 (i.e. it re-adds in some functions that were deleted).
As a side bonus, now the Gestures CTRL+G menu option shows as being checked when the gestures floater is active.