Fix for EXT-5120 "Mouse-up in the scrollbar to the right of a link in nearby chat window opens the link".
Added an optional bool parameter "hit_past_end_of_line" to LLTextBase:: getDocIndexFromLocalCoord() and getSegmentAtLocalPos().  Default is true, which gives the old behavior.  If false, clicks past the end-of-line won't map to the last segment on the line.

Made LLNormalTextSegment::handleHover(), handleRightMouseDown(), handleMouseDown(), and handleMouseUp() call getSegmentAtLocalPos() in this new mode before processing URL actions, and only do the processing if it returns the same segment.

Reviewed by Richard.