Fix for EXT-6287 (fullscreen flash video plays behind the viewer instead of in front)
Made SLPlugin into a bundled app on the Mac (this is apparently necessary for the plugin's window to be allowed to get focus).

Changed the Info.plist key SLPlugin uses to keep itself out of the dock from LSBackgroundOnly to LSUIElement (this allows it to get focus and display UI).

Added some Mac-specific code to slplugin.cpp to manipulate window layers and bring the plugin process to the foreground when something in the process opens a window, and to bring the viewer to the foreground when the last window in the process is closed.