Fixed critical bug EXT-6990 ( Unable to select group on 'People->My groups' tab)
This is workaroung to actually enable selection functionality in that tab. Copied default "allow_select" value from widget to group list xml.
Trivial change, not reviewed.


Real reason is that for some reason base attributes are not loaded from the flat_list_view.xml file.
It is investigated in scope of EXT-6990.