EXT-7901 FIXED Allow clicking on http://secondlife.com/app URLs.
The SLE code makes any http://<host>/app, http://<host>/region,
or http://<host>/secondlife URL be treated as a SLurl rather than a
normal http URL. This seems to be overly general, and has bitten us
because http://secondlife.com/app is used as a redirect URL for server
release notes.

In the short-term, I've made an exception for hostnames that match
"secondlife.com", so that the server release notes work again.

In the long-term, we should assess whether treating any URL that starts
with a path of "app", "region", or "secondlife" should be treated as a
SLurl. What happens when google adds "http://www.google.com/app"?