EXT-7838 FIX - viewer 2.1 avatars look ruth-shaped to viewer-2 users
Added a third visual param group VISUAL_PARAM_GROUP_TWEAKABLE_NO_TRANSMIT
to specify visual params that are user-tweakable, but should not be sent
over the network.

We should have this group *only* for new user-tweakable visual parameters
that should not be sent over the network. These should only be used for
*new* parameters that only contribute to the generation of baked textures.

Code reviewed by Richard