EXT-8356 FIXED Enabled localization of the "Touch" item of object context menu.
Since label of the "Touch" object context menu item may be overridden by the object with
something like "Sit" or "Kiss", i.e. the label can be different from the default one,
there is a special code for handling this.
The code uses parameter of the on_enable callback as a label for the menu item.
So, the fix is to pass localized label to on_enable callback of the "Object Touch" menu item.

Reviewed by Mike Antipov at https://codereview.productengine.com/secondlife/r/759/