EXT-8372 FIXED Added possibility to localize some system device names.
All device names including "No Device" are get from the Vivox plugin. So, just common names are added to localize table.
 * "Default system device"
 * "No device"

"Microphone" is a part of device name and can not be localized without a hack.

NOTE: this code will only work on English OS. On other system locales devices will be shown on system language independent on viewer locale.

Reviewed by Vadim Savchuk at https://codereview.productengine.com/secondlife/r/767/