EXT-8343 FIXED Fixed inability to rename a wearable containing non-ASCII character in the name.
It is actually possible to edit a default item name in My Inventory: when
user clicks on the input field, all its text gets selected, so user can just
enter the new name (although editing the old name is disallowed).

This fix implements the same behavior in the Edit Wearable panel by
selecting the name when the input field gets focused.

Reviewed by Mike Antipov at https://codereview.productengine.com/secondlife/r/763/