More cleanup around llmediadataclient.
LLMediaDataClient::Request no longer stores the LLSD payload -- it now uses the link to the LLVOVolume (which it was already keeping around) to generate the LLSD for the request at the time it's sent.

LLMediaDataClient::Request no longer stores a capability name.  Instead it uses the reference to its parent LLMediaDataClient subclass to refer to it as needed.

LLMediaDataClient::Request is now a base class, with subclasses for each specific request type -- GET, UPDATE, and NAVIGATE.

Responders are now created by a virtual method in the LLMediaDataClient::Request subclass instead of being created by the MDC.  This allows different request types to use different responder classes.

Fixed an issue with LLMediaDataClient::QueueTimer and LLMediaDataClient::RetryTimer that caused the unit test to fail (they now do all the work in their tick() functions instead of using the destructor).