EXT-8534 FIXED Fixed the 'Wear' in My Outfits being disabled for outfits consisting of body parts only.
Changed the condition to enable the "Wear" button in My Outfits and the corresponding item in the outfit context menu.
They now get enabled for any outfit that isn't the base outfit and contains non-worn wearables.

By the way, did a minor cleanup: moved an LLAgentWearables.isCOFChangeInProgress() call to LLAppearanceMgr::getCanAddToCOF() to avoid code duplication.

Reviewed by Sergey Litovchuk at https://codereview.productengine.com/secondlife/r/832/