BUILDFIX: viewer-mesh mac build
xcode didn't like the way we were defining inner classes and assuming that if the outer class
was a friend that the inner class would be too. Couldn't forward-declare the inner class
so had to re-arrange the order of class declaration.

Between re-ordering the class declarations, and explicitly making the inner class a friend,
this un-sticks the mac build, and makes it happy.

Code reviewed by davep.