STORM-311 FIXED "Share" button in My Inventory SP was not updated on Current Outfit changes.
When you wear an inventory item, the code that disables/enables the button was called as soon
as the item got linked to COF, before it actually appeared on your avatar. However
to determine whether to enable the button, the code checked avatar appearence.

I fixed it to take the COF link into account, i.e. to treat items linked to COF as
worn (=not shareable), no matter has appearance been updated or not.