STORM-1352 WIP Attempting to fix a crash in LLNearbyChatScreenChannel::showToastsBottom().
Apparently, a nearby chat toast got somehow destroyed while still remaining in the list of active toasts.
Attempt to sort active toasts in showToastsBottom() then triggered the crash.

I don't know how to reproduce the crash, i.e. force destroying a toast in a way that its onClose() method (which would remove references to the toast) isn't called.
So we'll just remove references to the toast whenever it's destroyed.