STORM-1600 FIXED Folder view stole "Delete" key presses from the line editor when renaming a newly created inventory item.
Reason:
LLFocusMgr::setKeyboardFocus() which is called from LLLineEditor::setFocus()
makes the folder view the edit menu handler, thus it receives the Delete key
presses instead of the line editor.

Fix:
Make sure the line editor becomes the edit menu handler whenever it's focused,
no matter is it a child of a folder view or not.