EXP-1489 FIXED (Cannot build notifications not being shown when chat floater closed with chat log toggled open)
Reason: There is a case when the chat floater is minimized while its visibility = true as well as its chat panel visibility = true. In this case notification won't be shown to the user. It will be shown in chat history which is in minimized floater.
Solution: Check for a floater minimized state.